+++ src/main/java/com/hta2405/unite/dto/Emp.java
@@ -36,6 +36,6 @@ public class Emp {
     private boolean married;
     private boolean child;
     private String etype;
-    private int vacationCount;
+    private Long vacationCount;
     private boolean hired;
 }
